requirements.txt
streamlit==1.46.0
tensorflow==2.18.0
keras==3.8.0
numpy==2.0.2
pandas==2.2.2
nltk==3.9.1
matplotlib==3.10.0
seaborn==0.13.2
google-api-python-client==2.172.0
python-dotenv==1.1.1
